Commenting out BitHub scores for now
I don't think it makes sense right now to include the post's score from BitHub without the ability to also vote up or at least view that content on Bithub. All the links point to the actual content (on the forum, not Bithub, for example).

I think we should add the score back ASAP, when we can also direct people to login to Bithub and vote. But I don't want this to hold us up from launching a 1.0

Signed-off-by: Tom Greever <anybody@tomgreever.com>